pyproject.toml
[project]
name = "poly-sports"
version = "0.1.0"
description = "A Python package for fetching and analyzing sports market data from Polymarket"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "py-clob-client",
    "requests",
    "python-dotenv",
    "pandas",
    "pytest",
    "pytest-mock",
    "rapidfuzz",
    "python-dateutil",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["poly_sports"]

